StateCoordinator
Installation Files for State Coordinators have been modified and updated. The filenames are now names in the way they need to be installed. The files that reside in Scripts\AIO and Scripts\Laptops are the final renditions. Process is now to run the files from the USB folder after having copied the SC2022 folder to c: drive